reafactor: factor a component into a separate file
